add a sample project with a graph test
A sample project has been added, that can be run.

This project can also be tested.

One of the tests runs the entire pypeman graph and checks
messages went through
add a sample project with a graph test

A sample project has been added, that can be run.

This project can also be tested.

One of the tests runs the entire pypeman graph and checks
messages went through